U3 Unit Project.
I tried to get the bullet code to work, but I had to scrap it because I
couldn’t get the x and y of the player. Enemies were also very hard to
do. I managed to get box 2d to work well. Not my finest work, but I’m
happy with how it turned out.